requirements/PartCrafter.txt
--extra-index-url https://download.pytorch.org/whl/rocm6.3
accelerate==1.9.0
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
blessed==1.21.0
bokeh==3.7.3
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colormaps==0.4.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
deepspeed==0.17.2
diffusers==0.34.0
einops==0.8.1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.13.1
fonttools==4.59.0
freetype-py==2.5.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
gpustat==1.1.1
gradio==5.38.2
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
jaxtyping==0.3.2
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.0.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
nvidia-ml-py==12.575.51
nvitop==1.5.2
omegaconf==2.3.0
opencv-python==4.11.0.86
orjson==3.11.1
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.0.0
platformdirs==4.3.8
plotly==6.2.0
proglog==0.1.12
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
Pygments==2.19.2
PyOpenGL==3.1.0
pyparsing==3.2.3
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-triton-rocm==3.3.1
pytz==2025.2
PyYAML==6.0.2
rdkit==2025.3.3
regex==2024.11.6
requests==2.32.4
rich==14.1.0
ruff==0.12.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentry-sdk==2.33.2
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
starlette==0.47.2
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1+rocm6.3
torchaudio==2.7.1+rocm6.3
torchvision==0.22.1+rocm6.3
tornado==6.5.1
tqdm==4.67.1
transformers==4.54.0
trimesh==4.7.1
triton==3.3.1
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wadler_lindig==0.1.7
wandb==0.21.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
xyzservices==2025.4.0
zipp==3.23.0